Improved print_logo function
Use array of integers instead of strings to increase performance
and to put my mind at ease. Improved logo look to be more in
line with refactored character. There isn't a great way of doing
pixel art in BASH as storring all of the RGB codes would be large
and harder to read.